+++ src/modules/web-inbox/WebInbox.js
@@ -12,6 +12,7 @@ export class Inbox extends HTMLElement {
   }
 
   isInboxOpen = false
+  isInboxFromFlutter = false
   selectedCategory = null
   unviewedMessages = {}
   unviewedCounter = 0
@@ -42,15 +43,18 @@ export class Inbox extends HTMLElement {
   }
 
   set incomingMessagesForPreview (msgs = []) {
+    const previewMsgs = {}
     if (msgs.length > 0 && this.inbox) {
       this.isPreview = true
       this.unviewedCounter = 0
       msgs.forEach((m) => {
-        m.id = `${m.wzrk_id.split('_')[0]}_${Date.now()}`
-        this.unviewedMessages[m.id] = m
+        const key = `${m.wzrk_id.split('_')[0]}_${Date.now()}`
+        m.id = key
+        previewMsgs[key] = m
+        this.unviewedMessages[key] = m
         this.unviewedCounter++
       })
-      this.buildUIForMessages(msgs)
+      this.buildUIForMessages(previewMsgs)
       this.updateUnviewedBadgeCounter()
     }
   }
@@ -220,12 +224,6 @@ export class Inbox extends HTMLElement {
     this.emptyInboxMsg.innerText = 'All messages will be displayed here.'
     this.inboxCard.appendChild(this.emptyInboxMsg)
 
-    if (this.config.hidePoweredByCT === false) {
-      const poweredByText = this.createEl('img', 'poweredByCT')
-      poweredByText.src = 'https://d2r1yp2w7bby2u.cloudfront.net/js/PB_CT_new.png'
-      this.inbox.appendChild(poweredByText)
-    }
-
     // Intersection observer for notification viewed
     const options = {
       root: this.inboxCard,
@@ -380,7 +378,11 @@ export class Inbox extends HTMLElement {
           }
         }
       } else if (this.inboxSelector.contains(e.target) || this.isInboxOpen) {
-        this.toggleInbox(e)
+        if (this.isInboxFromFlutter) {
+          this.isInboxFromFlutter = false
+        } else {
+          this.toggleInbox(e)
+        }
       }
     }
   })()
@@ -423,6 +425,7 @@ export class Inbox extends HTMLElement {
   // create a separte fn fro refactoring
   toggleInbox (e) {
     this.isInboxOpen = !this.isInboxOpen
+    this.isInboxFromFlutter = !!e?.rect
     if (this.isInboxOpen) {
       this.inboxCard.scrollTop = 0
       !this.isPreview && this.deleteExpiredAndGetUnexpiredMsgs()
@@ -494,7 +497,6 @@ export class Inbox extends HTMLElement {
   getInboxStyles () {
     const headerHeight = 36
     const categoriesHeight = this.config.categories.length ? 64 : 16
-    const hidePoweredByCTHeight = this.config.hidePoweredByCT === false ? 32 : 0
 
     const styles = {
       panelBackgroundColor: this.config.styles.panelBackgroundColor,
@@ -506,7 +508,7 @@ export class Inbox extends HTMLElement {
       categoriesTitleColor: this.config.styles.categories.titleColor,
       selectedCategoryTabColor: this.config.styles.categories.selectedTab.tabColor,
       selectedCategoryTitleColor: this.config.styles.categories.selectedTab.titleColor,
-      headerCategoryAndPoweredByCTHeight: headerHeight + categoriesHeight + hidePoweredByCTHeight
+      headerCategoryHeight: headerHeight + categoriesHeight
     }
     if (this.config.styles.categories.borderColor) {
       styles.categoriesBorderColor = this.config.styles.categories.borderColor