+++ src/util/tr.js
@@ -39,8 +39,7 @@ const _tr = (msg, {
   device,
   session,
   request,
-  logger,
-  isWebPopUpSpamControlDisabled
+  logger
 }) => {
   const _device = device
   const _session = session
@@ -132,7 +131,7 @@ const _tr = (msg, {
         const campaignSessionCount = sessionObj[campaignId]
         const totalSessionCount = sessionObj.tc
         // dnd
-        if (campaignSessionCount === 'dnd' && !isWebPopUpSpamControlDisabled) {
+        if (campaignSessionCount === 'dnd' && !$ct.dismissSpamControl) {
           return false
         }
 
@@ -358,11 +357,12 @@ const _tr = (msg, {
       if (doCampHouseKeeping(targetingMsgJson) === false) {
         return
       }
-      if (isWebPopUpSpamControlDisabled && document.getElementById(divId) != null) {
+      if ($ct.dismissSpamControl && document.getElementById(divId) != null) {
         const element = document.getElementById(divId)
         element.remove()
       }
-      if (document.getElementById(divId) != null) {
+      // ImageOnly campaign and Interstitial/Exit Intent shouldn't coexist
+      if (document.getElementById(divId) != null || document.getElementById('intentPreview') != null) {
         return
       }
       const msgDiv = document.createElement('div')
@@ -380,13 +380,14 @@ const _tr = (msg, {
 
     const divId = 'wizParDiv' + displayObj.layout
 
-    if (isWebPopUpSpamControlDisabled && document.getElementById(divId) != null) {
+    if ($ct.dismissSpamControl && document.getElementById(divId) != null) {
       const element = document.getElementById(divId)
       element.remove()
     }
     if (document.getElementById(divId) != null) {
       return
     }
+
     $ct.campaignDivMap[campaignId] = divId
     const isBanner = displayObj.layout === 2
     const msgDiv = document.createElement('div')
@@ -500,6 +501,10 @@ const _tr = (msg, {
     msgDiv.appendChild(iframe)
     const ifrm = (iframe.contentWindow) ? iframe.contentWindow : (iframe.contentDocument.document) ? iframe.contentDocument.document : iframe.contentDocument
     const doc = ifrm.document
+
+    // Dispatch event for popup box/banner close
+    const closeCampaign = new Event('CT_campaign_rendered')
+    document.dispatchEvent(closeCampaign)
 
     doc.open()
     doc.write(html)
@@ -701,13 +706,13 @@ const _tr = (msg, {
       targetingMsgJson = targetObj
     }
 
-    if (isWebPopUpSpamControlDisabled && targetingMsgJson.display.wtarget_type === 0 && document.getElementById('intentPreview') != null && document.getElementById('intentOpacityDiv') != null) {
+    if ($ct.dismissSpamControl && targetingMsgJson.display.wtarget_type === 0 && document.getElementById('intentPreview') != null && document.getElementById('intentOpacityDiv') != null) {
       const element = document.getElementById('intentPreview')
       element.remove()
       document.getElementById('intentOpacityDiv').remove()
     }
-
-    if (document.getElementById('intentPreview') != null) {
+    // ImageOnly campaign and Interstitial/Exit Intent shouldn't coexist
+    if (document.getElementById('intentPreview') != null || document.getElementById('wzrkImageOnlyDiv') != null) {
       return
     }
     // dont show exit intent on tablet/mobile - only on desktop
@@ -720,12 +725,15 @@ const _tr = (msg, {
     if (doCampHouseKeeping(targetingMsgJson) === false) {
       return
     }
+
     const campaignId = targetingMsgJson.wzrk_id.split('_')[0]
     $ct.campaignDivMap[campaignId] = 'intentPreview'
     let legacy = false
     const opacityDiv = document.createElement('div')
     opacityDiv.id = 'intentOpacityDiv'
-    opacityDiv.setAttribute('style', 'position: fixed;top: 0;bottom: 0;left: 0;width: 100%;height: 100%;z-index: 2147483646;background: rgba(0,0,0,0.7);')
+    const opacity = targetingMsgJson.display.opacity || 0.7
+    const rgbaColor = `rgba(0,0,0,${opacity})`
+    opacityDiv.setAttribute('style', `position: fixed;top: 0;bottom: 0;left: 0;width: 100%;height: 100%;z-index: 2147483646;background: ${rgbaColor};`)
     document.body.appendChild(opacityDiv)
 
     const msgDiv = document.createElement('div')
@@ -809,6 +817,10 @@ const _tr = (msg, {
     msgDiv.appendChild(iframe)
     const ifrm = (iframe.contentWindow) ? iframe.contentWindow : (iframe.contentDocument.document) ? iframe.contentDocument.document : iframe.contentDocument
     const doc = ifrm.document
+
+    // Dispatch event for interstitial/exit intent close
+    const closeCampaign = new Event('CT_campaign_rendered')
+    document.dispatchEvent(closeCampaign)
 
     doc.open()
     doc.write(html)